pyproject.toml
[tool.poetry]
name = "fastapi_backend"
version = "0.1.1"
description = "Test"
authors = ["Rishi Latchmepersad <latchmepersad@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.111.1"
uvicorn = "^0.30.3"
sqlalchemy = "^2.0.31"
psycopg2-binary = "^2.9.9"
python-dotenv = "^1.0.1"
fastapi-sqlalchemy = "^0.2.1"
alembic = "^1.13.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
